+++ src/blockchain.h
@@ -5,18 +5,17 @@
 extern "C" {
 #endif
 
-#include <stddef.h>
-#include <stdint.h>
+#include "precompiled_header.h"
 
 #include "block.h"
 
 typedef struct 
 {
-    uint32_t difficulty;        // The difficulty of the mining process (amount of leading 0h halfbytes -> 0000, in the ownhash)
+    uint8_t difficulty;         // The difficulty of the mining process (amount of leading 0h halfbytes -> 0000, in the ownhash)
     size_t allocatedBlocks;     // The amount of allicated bytes of the dynamic "blocks" array
     size_t usedBlocks;          // The amount of used bytes in the dynamic "blocks" array
     block_t * blocks;           // The blocks that the blockchain contains (dynamic array)
-}block_chain_t;
+} block_chain_t;
 
 /// @brief Adds a new block to the blockchain
 /// @param blockChain The blockchain where a new block is added
@@ -32,7 +31,7 @@ void blockchain_free(block_chain_t blockChain);
 /// @param blockChain The blockchain that is initialized
 /// @param difficulty The difficulty that the blockchain uses
 /// @return 0 if sucessful
-int blockchain_init(block_chain_t * blockChain, uint32_t difficulty);
+int blockchain_init(block_chain_t * blockChain, uint8_t difficulty);
 
 /// @brief Prints the blocks that the blockchain contains
 /// @param blockChain The blockchain that is printed
